+++ Core/World/Geometry/Builder/DoomGeometryBuilder.cs
@@ -9,6 +9,7 @@ using Helion.Maps.Specials.Vanilla;
 using Helion.Maps.Specials.ZDoom;
 using Helion.Resources;
 using Helion.Util;
+using Helion.Util.Extensions;
 using Helion.Util.Loggers;
 using Helion.World.Bsp;
 using Helion.World.Geometry.Lines;